backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==1.4.50
psycopg2-binary==2.9.7
python-dotenv==1.0.1
alembic==1.13.1
pydantic==2.6.3
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic-settings==2.1.0 